feat: convert bigvalue
